+++ utils/constants.js
@@ -5,12 +5,8 @@
 /**
  * Template types for form generation
  */
-const TEMPLATE_TYPES = {
+export const TEMPLATE_TYPES = {
   WITH_1_PAGE: 'WITH_1_PAGE',
   WITH_4_PAGES: 'WITH_4_PAGES',
   FORM_ENGINE: 'FORM_ENGINE',
-};
-
-module.exports = {
-  TEMPLATE_TYPES,
 };